feat: reorganize routes
