+++ src/views/dashboard/InfrastructureView.vue
